Issue #1718, #1727 Surelog parsing is pretty slow
Resolution
Replacing use of standard RTTI i.e. use of dynamic_cast
with a custom implementation that avoid having to enable
RTTI across the entire application.

This change replaces all uses of dynamic_cast with
<base_class>_cast function and updates the code generator
to do the same.

It also reverts the following past change which was another
approach to solving the performance but much narrower.

  Revert "Optimizing Cpp runtime"

  This reverts commit 94f364e4f93b6583bd883614577b252568430bb4.